+++ CHANGELOG.md
@@ -16,11 +16,15 @@ Most recent releases are shown at the top. Each release shows:
   a `.vec.gz` file from [here](https://fasttext.cc/docs/en/crawl-vectors.html). The URL (or path) is
   supplied as `wv_path_or_url` argument to data-loading functions like `entities_from_array` and `entities_from_txt`
 - `show_random_images`: show random images from folder in Jupyter notebook
+- `NERPreprocessor` now includes a `preprocess_test` method for easier evaluation of test sets in datasets
+   that contain a training, validation, and test set
 
 ### Changed:
 - ensure `DISABLE_V2_BEHAVIOR=True` when `ImagePredictor.explain` is invoked
 - added `SUPPRESS_TF_WARNINGS` environment variable.  Default is '1'. If set to '0', TF warnings will be displayed.
-- `merge_entities` method of `ktrain.text.shallownlp.ner.NER` changed to `merge_tokens`
+- `merge_entities` method of `ktrain.text.shallownlp.ner.NER` changed to `merge_tokens` 
+- moved `load_predictor` to constructor in `krain.text.shallownlp.ner.NER`
+- `ktrain.text.shallownlp.ner.NER` now supports `predictor_path` argument
 
 ### Fixed:
 - convert `class_names` to strings in `core.validate` to prevent error from scikit-learn